Time: 28 ms (24.15%) | Memory: 57 MB (49.43%) - LeetSync

+++ 3094-minimum-number-of-operations-to-make-array-empty/minimum-number-of-operations-to-make-array-empty.java
@@ -15,7 +15,7 @@ class Solution {
             if(freq.get(key)==1) return -1;
             else {
                 // System.out.println(key + " " + freq.get(key) + "; " + Math.ceil((double)freq.get(key)/3));
-                ans += (int)Math.ceil((double)freq.get(key)/3);
+                ans += Math.ceil((double)freq.get(key)/3);
             }
         }
         return ans;
